refactor: 파일 구조 변경 및 탭 초기값 문제 수정 (#18)
