+++ src/app/health/layout.jsx
@@ -1,7 +1,7 @@
 "use client";
 
-import { SelectedPetProvider } from "../context/SelectedPetContext";
-import Header from "../activity/components/ActivityHeader";
+import { SelectedPetProvider } from "./context/SelectedPetContext";
+import Header from "./activity/components/ActivityHeader";
 
 export default function HealthLayout({ children }) {
   return (
